give area() the same behaviour as area_by_strata()
Gives the option to return the area that was mapped, the estimated true
area, or the area that was mapped correctly for a given class.

Makes returning the estimated true area the default for both area() and
area_by_strata().

Adds better documentation of the different behaviours for the area
classes.